make js more modular
